backend/pyproject.toml
[project]
name = "learn-backend"
version = "0.1.0"
description = "1111 - Generative Learning Platform Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic-ai-slim[anthropic]>=0.1.0",
    "anthropic>=0.40.0",
    "sse-starlette>=2.0.0",
    "pyjwt>=2.0.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
